Refactor tests, remove uopz dependency
