+++ composer.json
@@ -29,19 +29,18 @@
     ],
     "require": {
         "php": "^8.1",
-        "yiisoft/rbac": "^2.0",
+        "yiisoft/rbac": "^2.1",
         "yiisoft/var-dumper": "^1.7"
     },
     "require-dev": {
-        "ext-uopz": "*",
-        "maglnet/composer-require-checker": "^4.3",
-        "phpunit/phpunit": "^10.5.2",
-        "rector/rector": "^2.0.2",
-        "roave/infection-static-analysis-plugin": "^1.18",
-        "slope-it/clock-mock": "0.5.0",
-        "spatie/phpunit-watcher": "^1.23",
-        "vimeo/psalm": "^4.30|^5.17",
-        "yiisoft/files": "^1.0"
+        "maglnet/composer-require-checker": "^4.7.1",
+        "phpunit/phpunit": "^10.5.45",
+        "psr/clock": "^1.0",
+        "rector/rector": "^2.0.11",
+        "roave/infection-static-analysis-plugin": "^1.35",
+        "spatie/phpunit-watcher": "^1.24",
+        "vimeo/psalm": "^5.26.1",
+        "yiisoft/files": "^1.0.2"
     },
     "autoload": {
         "psr-4": {
